Add Category class
Phran needs a way to categorize people who are volunteering for FFF. I devised a system that will help Phran keep track of volunteers' responsibilities through the admin console.

Phran can assign a category to any person in the admin console, search the people in a category, and then download a CSV of the those people in that category.

	- add category to user

	- export CSV of users with categories